+++ framework/console/chi_console.cc
@@ -1,18 +1,319 @@
 #include "console/chi_console.h"
-
+#include "lua/chi_modules_lua.h"
+#include "chi_configuration.h"
+#include "ChiObjectFactory.h"
 #include "chi_runtime.h"
 #include "chi_log.h"
-
+#include "chi_log_exceptions.h"
+#include "chi_lua.h"
+#include "chi_mpi.h"
+#include "chi_console_structs.h"
 #include "chi_utils.h"
-
-#include "ChiObjectFactory.h"
+#if defined(__MACH__)
+#include <mach/mach.h>
+#else
+#include <unistd.h>
+#endif
+#include <iostream>
 
 namespace chi::lua_utils
 {
 int chiMakeObject(lua_State* L);
 }
 
-// #############################################################################
+namespace chi
+{
+
+RegisterLuaFunction(Console::LuaWrapperCall, chi_console, LuaWrapperCall);
+
+Console&
+Console::GetInstance() noexcept
+{
+  static Console singleton;
+  return singleton;
+}
+
+Console::Console() noexcept : console_state_(luaL_newstate())
+{
+}
+
+void
+Console::LoadRegisteredLuaItems()
+{
+  //=================================== Initializing console
+  auto& L = GetConsoleState();
+
+  luaL_openlibs(L);
+
+  //=================================== Register version
+  lua_pushstring(L, PROJECT_VERSION);
+  lua_setglobal(L, "chi_version");
+  lua_pushinteger(L, PROJECT_MAJOR_VERSION);
+  lua_setglobal(L, "chi_major_version");
+  lua_pushinteger(L, PROJECT_MINOR_VERSION);
+  lua_setglobal(L, "chi_minor_version");
+  lua_pushinteger(L, PROJECT_PATCH_VERSION);
+  lua_setglobal(L, "chi_patch_version");
+
+  //=================================== Registering functions
+  chi_modules::lua_utils::RegisterLuaEntities(L);
+
+  //=================================== Registering static-registration
+  //                                    lua functions
+  for (const auto& [key, entry] : lua_function_registry_)
+    SetLuaFuncNamespaceTableStructure(key, entry.function_ptr);
+
+  //=================================== Registering LuaFunctionWrappers
+  for (const auto& [key, entry] : function_wrapper_registry_)
+    if (entry.call_func) SetLuaFuncWrapperNamespaceTableStructure(key);
+
+  for (const auto& [key, value] : lua_constants_registry_)
+    SetLuaConstant(key, value);
+
+  //=================================== Registering solver-function
+  //                                    scope resolution tables
+  const auto& object_maker = ChiObjectFactory::GetInstance();
+  for (const auto& entry : object_maker.Registry())
+    SetObjectNamespaceTableStructure(entry.first);
+}
+
+void
+Console::FlushConsole()
+{
+  try
+  {
+    for (auto& command : command_buffer_)
+    {
+      bool error = luaL_dostring(console_state_, command.c_str());
+      if (error)
+      {
+        Chi::log.LogAll() << lua_tostring(console_state_, -1);
+        lua_pop(console_state_, 1);
+      }
+    }
+  }
+  catch (const std::exception& e)
+  {
+    Chi::log.LogAllError() << e.what();
+    Chi::Exit(EXIT_FAILURE);
+  }
+}
+
+int
+Console::LuaWrapperCall(lua_State* L)
+{
+  const int num_args = lua_gettop(L);
+  // We do not check for the required parameters here because we want
+  // to make this function call as fast as possible. Besides, via the
+  // static registration we should never run into an issue here.
+
+  auto& console = Console::GetInstance();
+
+  const auto& registry = console.function_wrapper_registry_;
+
+  const std::string fname = lua_tostring(L, 1);
+
+  ChiLogicalErrorIf(registry.count(fname) == 0,
+                    std::string("Wrapper with name \"") + fname + "\" not in console registry.");
+
+  const auto& reg_entry = registry.at(fname);
+
+  auto input_params = reg_entry.get_in_params_func();
+
+  ParameterBlock main_arguments_block;
+  for (int p = 2; p <= num_args; ++p)
+  {
+    const std::string arg_name = "arg" + std::to_string(p - 2);
+
+    if (lua_isboolean(L, p)) main_arguments_block.AddParameter(arg_name, lua_toboolean(L, p));
+    else if (lua_isinteger(L, p))
+      main_arguments_block.AddParameter(arg_name, lua_tointeger(L, p));
+    else if (lua_isnumber(L, p))
+      main_arguments_block.AddParameter(arg_name, lua_tonumber(L, p));
+    else if (lua_isstring(L, p))
+      main_arguments_block.AddParameter(arg_name, lua_tostring(L, p));
+    else if (lua_istable(L, p))
+    {
+      auto block = chi_lua::TableParserAsParameterBlock::ParseTable(L, p);
+      block.SetBlockName(arg_name);
+      std::string scope = fname + ":";
+      scope.append(arg_name + " ");
+      block.SetErrorOriginScope(scope);
+      main_arguments_block.AddParameter(block);
+    }
+    else
+      ChiInvalidArgument("In call to \"" + fname +
+                         "\": Unsupported argument "
+                         "type \"" +
+                         lua_typename(L, lua_type(L, p)) + "\" encountered.");
+  }
+  // Set input parameters here
+  input_params.SetErrorOriginScope(fname + "()");
+  input_params.AssignParameters(main_arguments_block);
+
+  auto output_params = reg_entry.call_func(input_params);
+
+  output_params.SetErrorOriginScope(fname + ":output:");
+  chi_lua::PushParameterBlock(L, output_params);
+
+  const int num_sub_params = static_cast<int>(output_params.NumParameters());
+
+  return output_params.IsScalar() ? 1 : num_sub_params;
+}
+
+void
+Console::RunConsoleLoop(char*) const
+{
+  Chi::log.Log() << "Console loop started. "
+                 << "Type \"exit\" to quit (or Ctl-C).";
+
+  /** Wrapper to an MPI_Bcast call for a single integer
+   * broadcast from location 0. */
+  auto BroadcastSingleInteger = [](int* int_being_bcast)
+  {
+    MPI_Bcast(int_being_bcast, // buffer
+              1,
+              MPI_INT,        // count + type
+              0,              // root
+              Chi::mpi.comm); // communicator
+  };
+
+  /** Wrapper to an MPI_Bcast call for an array of characters
+   * broadcast from location 0. */
+  auto HomeBroadcastStringAsRaw = [](std::string string_to_bcast, int length)
+  {
+    char* raw_string_to_bcast = string_to_bcast.data();
+    MPI_Bcast(raw_string_to_bcast, // buffer
+              length,
+              MPI_CHAR,       // count + type
+              0,              // root
+              Chi::mpi.comm); // communicator
+  };
+
+  /** Wrapper to an MPI_Bcast call for an array of characters
+   * broadcast from location 0. This call is for non-home locations. */
+  auto NonHomeBroadcastStringAsRaw = [](std::string& string_to_bcast, int length)
+  {
+    std::vector<char> raw_chars(length + 1, '\0');
+    MPI_Bcast(raw_chars.data(), // buffer
+              length,
+              MPI_CHAR,       // count + type
+              0,              // root
+              Chi::mpi.comm); // communicator
+
+    string_to_bcast = std::string(raw_chars.data());
+  };
+
+  /** Executes a string within the lua-console. */
+  auto LuaDoString = [this](const std::string& the_string)
+  {
+    bool error = luaL_dostring(console_state_, the_string.c_str());
+    if (error)
+    {
+      Chi::log.LogAll() << lua_tostring(console_state_, -1);
+      lua_pop(console_state_, 1);
+    }
+  };
+
+  auto ConsoleInputNumChars = [](const std::string& input)
+  {
+    int L = static_cast<int>(input.size());
+    if (input == std::string("exit")) L = -1;
+
+    return L;
+  };
+
+  const bool HOME = Chi::mpi.location_id == 0;
+
+  while (not Chi::run_time::termination_posted_)
+  {
+    std::string console_input;
+
+    if (HOME) std::cin >> console_input; // Home will be waiting here
+
+    int console_input_len = ConsoleInputNumChars(console_input);
+
+    BroadcastSingleInteger(&console_input_len); // Non-Home locs wait here
+
+    if (console_input_len < 0) break;
+    else if (HOME)
+      HomeBroadcastStringAsRaw(console_input, console_input_len);
+    else
+      NonHomeBroadcastStringAsRaw(console_input, console_input_len);
+
+    try
+    {
+      LuaDoString(console_input);
+    }
+    catch (const Chi::RecoverableException& e)
+    {
+      Chi::log.LogAllError() << e.what();
+    }
+    catch (const std::exception& e)
+    {
+      Chi::log.LogAllError() << e.what();
+      Chi::Exit(EXIT_FAILURE);
+    }
+  } // while not termination posted
+
+  Chi::run_time::termination_posted_ = true;
+
+  Chi::log.Log() << "Console loop stopped successfully.";
+}
+
+CSTMemory
+Console::GetMemoryUsage()
+{
+  double mem = 0.0;
+#if defined(__MACH__)
+  struct mach_task_basic_info info;
+  mach_msg_type_number_t count = MACH_TASK_BASIC_INFO_COUNT;
+  long long int bytes;
+  if (task_info(mach_task_self(), MACH_TASK_BASIC_INFO, (task_info_t)&info, &count) != KERN_SUCCESS)
+  {
+    bytes = 0;
+  }
+  bytes = info.resident_size;
+  mem = (double)bytes;
+#else
+  long long int llmem = 0;
+  long long int rss = 0;
+
+  std::string ignore;
+  std::ifstream ifs("/proc/self/stat", std::ios_base::in);
+  ifs >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >>
+    ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >>
+    ignore >> ignore >> ignore >> ignore >> llmem >> rss;
+
+  long long int page_size_bytes = sysconf(_SC_PAGE_SIZE);
+  mem = rss * page_size_bytes;
+  /*
+  FILE* fp = NULL;
+  if((fp = fopen( "/proc/self/statm", "r" )) == NULL)
+    return 0;
+  if(fscanf(fp, "%*s%*s%*s%*s%*s%lld", &llmem) != 1)
+  {
+    fclose(fp);
+    return 0;
+  }
+  fclose(fp);*/
+
+  // mem = llmem * (long long int)sysconf(_SC_PAGESIZE);
+#endif
+
+  CSTMemory mem_struct(mem);
+
+  return mem_struct;
+}
+
+double
+Console::GetMemoryUsageInMB()
+{
+  CSTMemory mem_struct = GetMemoryUsage();
+
+  return mem_struct.memory_mbytes;
+}
+
 // Execute file
 /** Executes the given file in the Lua engine.
 \author Jan*/
@@ -404,11 +705,8 @@ chi::Console::DumpRegister() const
   Chi::log.Log() << "\n\n";
 }
 
-// ##################################################################
-/**Given an old status, will update the bindings for only newly registered
- * items.*/
 void
-chi::Console::UpdateConsoleBindings(const chi::RegistryStatuses& old_statuses)
+Console::UpdateConsoleBindings(const chi::RegistryStatuses& old_statuses)
 {
   auto ListHasValue = [](const std::vector<std::string>& list, const std::string& value)
   { return std::find(list.begin(), list.end(), value) != list.end(); };
@@ -425,3 +723,5 @@ chi::Console::UpdateConsoleBindings(const chi::RegistryStatuses& old_statuses)
     if (not ListHasValue(old_statuses.objfactory_keys_, key))
       if (entry.call_func) SetLuaFuncWrapperNamespaceTableStructure(key);
 }
+
+} // namespace chi
